Better log messages about disks
